Consolidate GSD workflow seams after code-quality review.
Extract shared projection flush, DB reopen, slice-research resolution, and artifact verification helpers so dispatch, prompts, recovery, and completion tools reuse canonical paths instead of duplicating logic.
